# src/api/openapi.yaml
openapi: 3.1.0
info:
  title: SAVED4 OpenAPI Spec
  description: API specification for SAVED4
  version: 0.1.0
  x-fiddle-import-file: true
paths:
...
  /companies/{username}:
    get:
      operationId: getCompany
      summary: Get company
      parameters:
        - $ref: "#/components/parameters/username"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      description: "Retrieves a company resource from the system by its username. If
        the company does not exist, a 404 response is returned. "
      tags:
        - companies
      security: []
    delete:
      operationId: deleteCompany
      summary: Delete company
      parameters:
        - $ref: "#/components/parameters/username"
      responses:
        "204":
          description: The request was successful, but there is no content to return in
            the response.
          content:
            application/json:
              schema:
                type: "null"
                title: DeleteCompanyNoContent
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      description: Deletes the company identified by the specified username.
      tags:
        - companies
      security: []
    patch:
      operationId: updateAnExistingCompany
      summary: Update an existing company
      parameters:
        - $ref: "#/components/parameters/username"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      description: Updates the details of an existing company identified by the
        username in the path.
      tags:
        - companies
      security: []
      requestBody:
        required: true
        description: A JSON object containing the fields to be updated.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyPartialUpdate"
...
  /routes/{route_id}:
    get:
      summary: Get route
      operationId: getRoute
      description: Returns a route identified by route id.
      tags:
        - routes
      security: []
      parameters:
        - $ref: "#/components/parameters/route_id"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
    patch:
      summary: Update route
      operationId: updateRoute
      description: Updates a route.
      tags:
        - routes
      security: []
      parameters:
        - $ref: "#/components/parameters/route_id"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      requestBody:
        required: true
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutePartialUpdate"
    delete:
      summary: Delete route
      operationId: deleteRoute
      description: Deletes a route.
      tags:
        - routes
      security: []
      parameters:
        - $ref: "#/components/parameters/route_id"
      responses:
        "204":
          description: The request was successful, but there is no content to return in
            the response.
          content:
            application/json:
              schema:
                type: "null"
                title: DeleteRouteNoContent
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
  /routes:
    get:
      summary: List route
      operationId: listRoute
      description: Returns a list of all routes.
      tags:
        - routes
      security: []
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
                title: ListRouteOk
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      parameters:
        - in: query
          name: limit
          required: false
          style: form
          description: Maximum number of routes to return
          explode: true
          schema:
            type: integer
          allowReserved: false
    post:
      summary: Create route
      operationId: createRoute
      description: Creates a new route.
      tags:
        - routes
      security: []
      responses:
        "201":
          description: The request was successful, and a new resource was created.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteCreate"
                title: CreateRouteCreated
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      parameters: []
      requestBody:
        required: true
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteCreate"
...
components:
  parameters:
    username:
      name: username
      in: path
      required: true
      schema:
        type: string
...
    route_id:
      name: route_id
      in: path
      required: true
      schema:
        type: integer
  responses:
    BadRequest:
      description: The server could not understand the request due to invalid syntax.
        The client should modify the request and try again.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: array
                items:
                  type: object
                  properties:
                    message:
                      type: string
                  required:
                    - message
            required:
              - message
    Unauthorized:
      description: Authentication is required to access the requested resource. The
        client must include the appropriate credentials.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    Forbidden:
      description: The server understood the request, but refuses to authorize it.
        Ensure the client has appropriate permissions.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    InternalServerErrror:
      description: The server encountered an unexpected condition that prevented it
        from fulfilling the request. Report the issue to the support team if it
        persists.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    NotFound:
      description: The server cannot find the requested resource. The endpoint may be
        invalid or the resource may no longer exist.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
...
    Conflict:
      description: The request could not be completed due to a conflict with the
        current state of the resource. Resolve the conflict and try again.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
...
  securitySchemes:
    BearerJwt:
      type: http
      scheme: Bearer
      bearerFormat: Json Web Token (JWT)
      description: "Jwt Auth: Authenticated requests contain a valid Json Web Token
        (JWT) as part of the `Authorization: Bearer <JWT>` header."
  schemas:
...
    Route:
      type: object
      properties:
        vehicle_id:
          type: integer
          description: ID of the vehicle assigned to the route.
          example: 42
        route:
          type: array
          description: List of coordinates forming the route.
          items:
            type: object
            properties:
              latitude:
                type: number
                description: Latitude coordinate in the route.
                example: 52.3791
              longitude:
                type: number
                description: Longitude coordinate in the route.
                example: 4.8994
            required:
              - latitude
              - longitude
        route_id:
          type: integer
          description: Unique identifier for the route.
          example: 301
      required:
        - route
        - route_id
      description: A route taken by a vehicle to fulfill an order
    RoutePartialUpdate:
      type: object
      properties:
        vehicle_id:
          type: integer
          description: Updated ID of the vehicle assigned to the route.
          example: 43
        route:
          type: array
          description: Updated list of coordinates in the route.
          items:
            type: object
            properties:
              latitude:
                type: number
                description: Latitude coordinate in the route.
                example: 51.9225
              longitude:
                type: number
                description: Longitude coordinate in the route.
                example: 4.4792
            required:
              - latitude
              - longitude
      required: []
      description: Data transfer object for partially updating an existing Route
        (PATCH operation).
      x-fiddle-dto-info:
        baseSchemaName: Route
    RouteCreate:
      type: object
      properties:
        vehicle_id:
          type: integer
          description: ID of the vehicle to which the route will be assigned.
          example: 45
        route:
          type: array
          description: List of coordinates defining the new route.
          items:
            type: object
            properties:
              latitude:
                type: number
                description: Latitude coordinate.
                example: 52.0705
              longitude:
                type: number
                description: Longitude coordinate.
                example: 4.3007
            required:
              - latitude
              - longitude
      required:
        - route
      description: Data transfer object for creating a new Route.
      x-fiddle-dto-info:
        baseSchemaName: Route
...
    CompanySummary:
      type: object
      properties:
        username:
          type: string
          example: companyA
        name:
          type: string
          example: A
        location:
          type: object
          properties:
            latitude:
              type: number
              example: 41.1
            longitude:
              type: number
              example: 41.1
          required:
            - latitude
            - longitude
        closing_time:
          type: string
          format: time
          example: 14:30:00
        opening_time:
          type: string
          format: time
          example: 14:30:00
      required:
        - username
        - name
        - location
        - closing_time
        - opening_time
      description: A simplified representation of a Company.
      x-fiddle-dto-info:
        baseSchemaName: Company
    CompanyPartialUpdate:
      type: object
      properties:
        username:
          type: string
          example: companyA
        password:
          type: string
        name:
          type: string
          example: A
        location:
          type: object
          properties:
            latitude:
              type: number
              example: 41.1
            longitude:
              type: number
              example: 41.1
          required:
            - latitude
            - longitude
        closing_time:
          type: string
          format: time
          example: 14:30:00
        opening_time:
          type: string
          format: time
          example: 14:30:00
      required: []
      description: Data transfer object for partially updating an existing Company
        (PATCH operation).
      x-fiddle-dto-info:
        baseSchemaName: Company
...