# src/backend/src/main/java/org/saved4/saved4/dto/Route.java
package org.saved4.saved4.dto;

public class Route {
    // NOTE:
    // This class will be used for further development of routing.
    // Currently, we have only implemented a basic routing algorithm so this class is not used.
}